Changed double to BigDecimal

+++ UTP-5.S_BIGDECCAL.JAVABEAN/src/zad1/operation/Operation.java
@@ -1,6 +1,8 @@
 package zad1.operation;
 
+import java.math.BigDecimal;
+
 public interface Operation {
-    double operation(double a, double b);
+    BigDecimal operation(BigDecimal a, BigDecimal b);
 }
